map solve

+++ map.hpp
@@ -6,7 +6,7 @@
 /*   By: honlee <honlee@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/28 04:20:32 by honlee            #+#    #+#             */
-/*   Updated: 2021/04/29 14:14:35 by honlee           ###   ########.fr       */
+/*   Updated: 2021/04/30 00:09:45 by honlee           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,6 +15,7 @@
 
 # include "BST.hpp"
 # include "utils.hpp"
+# include "map_iterator.hpp"
 # include <algorithm>
 # include <cstddef>
 # include <iostream>
@@ -28,7 +29,7 @@ namespace ft
 		private :
 			node<TK, TV, Compare>	*root;
 			unsigned int				noe;
-				
+			
 		public	:
 			typedef TK key_type;
 			typedef TV mapped_type;
@@ -47,38 +48,34 @@ namespace ft
 
 			typedef Alloc allocator_type;
 			
-
+			//c plusplus 홈페이지에 나와있는 그대로 가따붙임 pair 를 직접 비교해주는 클레스.
 			class value_compare
 			{
-			private:
-				key_compare _cmp;
-	
-				value_compare(key_compare c) : _cmp(c)
-				{
-				}
-	
-				value_compare &operator=(const value_compare &o);
-	
-			public:
-				typedef bool result_type;
-				typedef value_type first_argument_type;
-				typedef value_type second_argument_type;
-	
-				~value_compare()
-				{
-				}
-	
-				value_compare(const value_compare &o) : _cmp(o._cmp)
-				{
-				}
+				private :
+					Compare comp;
+					value_compare (Compare c) : comp(c){}
+				public	:
+					typedef bool result_type;
+  					typedef value_type first_argument_type;
+  					typedef value_type second_argument_type;
 
-				bool operator()(const value_type &x, const value_type &y) const
-				{
-					return _cmp(x.first, y.first);
-				}	
+					value_compare(){}
+
+					value_compare(const value_compare& origin) : comp(origin.comp) {}
+
+					~value_compare(){}
+
+					value_compare& operator=(const value_compare& origin)
+					{
+						comp = origin.comp;
+						return (*this);
+					}
+
+					bool operator() (const value_type& x, const value_type& y) const
+  					{
+    					return comp(x.first, y.first);
+					}
 			};
-
-
 
 			//////////////////////////////////////////////////////////////////
 			//						constructor start						//
@@ -86,8 +83,7 @@ namespace ft
 
 			//default
 			explicit map (const key_compare& comp = key_compare(), const allocator_type& alloc = allocator_type()) : root(NULL), noe(0)
-			{
-				//std::cout << "default constrcutor called" << std::endl;
+			{				
 				(void)comp;
 				(void)alloc;
 			}
@@ -95,31 +91,34 @@ namespace ft
 			//range
 			template <class InputIterator>
   			map (InputIterator first, InputIterator last, const key_compare& comp = key_compare(), const allocator_type& alloc = allocator_type()) : root(NULL), noe(0)
-			{
-				//std::cout << "range constrcutor called" << std::endl;
+			{				
 				(void)comp;
 				(void)alloc;
+
+				insert(first, last);			
 			}
 			
 			//copy
 			map (const map& x) : root(NULL), noe(0) 
 			{
 				//std::cout << "copy constrcutor called" << std::endl;
-				this->root = new node<TK, TV, Compare>(x.root);
+				this->root = new node<TK, TV, Compare>(* x.root);
 				this->noe = x.noe;
+
 			}
 
 			map<TK, TV, Compare, Alloc>& operator= (const map& x)
 			{
 				this->root->deleteAll(this->root); // 가지고 있던 모든 것 파괴
-				this->root = new node<TK, TV, Compare>(x.root);
+				this->root = new node<TK, TV, Compare>(* x.root);
 				this->noe = x.noe;
 				return (*this);
 			}
 
 			~map()
 			{
-				this->root->deleteAll(this->root);
+				if (noe > 0)
+					this->root->deleteAll(this->root);
 			}
 
 			//////////////////////////////////////////////////////////////////
@@ -132,42 +131,42 @@ namespace ft
 
 			iterator begin()
 			{
-				return (iterator(root->getLeftest(root)));
+				return (iterator(root->getLeftest(root), root));
 			}
 
 			const_iterator begin() const
 			{
-				return (const_iterator(root->getLeftest(root)));
+				return (const_iterator(root->getLeftest(root), root));
 			}
 
 			iterator end()
 			{
-				return (iterator(NULL));
+				return (iterator(NULL, root));
 			}
 
 			const_iterator end() const
 			{
-				return (const_iterator(NULL));
+				return (const_iterator(NULL, root));
 			}
 
 			reverse_iterator rbegin()
 			{
-				return (reverse_iterator(root->getRightest(root)));
+				return (reverse_iterator(root->getRightest(root), root));
 			}
 
 			const_reverse_iterator rbegin() const
 			{
-				return (const_reverse_iterator(root->getRightest(root)));
+				return (const_reverse_iterator(root->getRightest(root), root));
 			}
 
 			reverse_iterator rend()
 			{
-				return (reverse_iterator(NULL));
+				return (reverse_iterator(NULL, root));
 			}
 
 			const_reverse_iterator rend() const
 			{
-				return (const_reverse_iterator(NULL));
+				return (const_reverse_iterator(NULL, root));
 			}
 
 			//////////////////////////////////////////////////////////////////
@@ -216,7 +215,7 @@ namespace ft
 				{
 					if ((ret = root->find(root, key)) != NULL) // 찾았다 (존재한다.)
 					{
-						return (root->ip.second);
+						return (ret->ip.second);
 					}
 					else
 					{
@@ -242,25 +241,26 @@ namespace ft
 				{
 					noe++;
 					root = new node<TK, TV, Compare>(val.first, val.second);
-					return (pair<iterator, bool>(root, true));
+					return (pair<iterator, bool>(iterator(root, root), true));
 				}
 				else
 				{
 					if ((ret = root->find(root, val.first)) != NULL) // 찾았다 (존재한다.)
 					{
-						return (pair<iterator, bool>(ret, false));
+						return (pair<iterator, bool>(iterator(ret, root), false));
 					}
 					else
 					{
 						noe++;
 						ret = root->mergeInsert(root, val.first, val.second);
-						return (pair<iterator, bool>(ret, true));
+						return (pair<iterator, bool>(iterator(ret, root), true));
 					}
 				}
 			}
 
 			iterator insert (iterator position, const value_type& val)
 			{
+				(void)position;
 				return (insert(val).first);
 			}
 
@@ -286,6 +286,8 @@ namespace ft
 					{
 						root->deleteNode(&root, root, k);
 						noe--;
+						if (noe == 0)
+							root = NULL;
 						return (1);
 					}
 					else
@@ -303,6 +305,8 @@ namespace ft
 					iter_next++;
 					erase(iter);
 					iter = iter_next;
+					if (iter == last)
+						break ;
 				}
 			}
 
@@ -330,7 +334,6 @@ namespace ft
 			//						Modifiers end							//
 			//////////////////////////////////////////////////////////////////
 
-
 			//////////////////////////////////////////////////////////////////
 			//						Observers start							//
 			//////////////////////////////////////////////////////////////////
@@ -339,7 +342,10 @@ namespace ft
 				return (Compare());
 			}
 
-			
+			value_compare value_comp() const
+			{
+				return (value_compare());
+			}
 
 			//////////////////////////////////////////////////////////////////
 			//						Observers end							//
@@ -352,15 +358,67 @@ namespace ft
 			iterator find (const key_type& k)
 			{
 				if (noe == 0)
-					return (iterator(NULL));
-				return (iterator(root->find(root, k)));
+					return (iterator(NULL, root));
+				return (iterator(root->find(root, k), root));
 			}
 
 			const_iterator find (const key_type& k) const
 			{
 				if (noe == 0)
-					return (const_iterator(NULL));
-				return (const_iterator(root->find(root, k)));
+					return (const_iterator(NULL, root));
+				return (const_iterator( root->find(root, k), root ) );
+			}
+
+			size_type count (const key_type& k) const
+			{
+				if (noe == 0)
+					return (0);
+				if (root->find(root, k) != NULL)
+					return (1);
+				else
+					return (0);
+			}
+
+			iterator lower_bound (const key_type& k)
+			{
+				if (noe == 0)
+					return (iterator(NULL, root));
+				else
+					return (iterator(root->getLowerBound(root, k), root));
+			}
+
+			const_iterator lower_bound (const key_type& k) const
+			{
+				if (noe == 0)
+					return (const_iterator(NULL, root));
+				else
+					return (const_iterator(root->getLowerBound(root, k), root));
+			}
+
+			iterator upper_bound (const key_type& k)
+			{
+				if (noe == 0)
+					return (iterator(NULL, root));
+				else
+					return (iterator(root->getUpperBound(root, k), root));
+			}
+
+			const_iterator upper_bound (const key_type& k) const
+			{
+				if (noe == 0)
+					return (const_iterator(NULL, root));
+				else
+					return (const_iterator(root->getUpperBound(root, k), root));
+			}
+
+			pair<iterator,iterator>             equal_range (const key_type& k)
+			{
+				return (pair<iterator, iterator>(lower_bound(k), upper_bound(k)));
+			}
+
+			pair<const_iterator,const_iterator> equal_range (const key_type& k) const
+			{
+				return (pair<const_iterator, const_iterator>(lower_bound(k), upper_bound(k)));
 			}
 
 			//////////////////////////////////////////////////////////////////
